pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.7.1"
description = "Gendiff (Hexlet Python Project Level 2)"
authors = ["neihaoo <neihaoo@gmail.com>"]
repository = "https://github.com/neihaoo/python-project-lvl2"
classifiers = [
  "Topic :: Utilities"
]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.0"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
